:sparkles: Namespaces: Base Component
