+++ src/app.ts
@@ -3,36 +3,9 @@
 /// <reference path="state/project-state.ts" />
 /// <reference path="util/validation.ts" />
 /// <reference path="decorators/autobind.ts" />
+/// <reference path="components/base-component.ts" />
 
 namespace App {
-  // Component Base Class
-  abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-    templateElement: HTMLTemplateElement;
-    hostElement: T;
-    element: U;
-
-    constructor(templateId: string, hostElementId: string, insertAtStart: boolean, newElementId?: string) {
-      this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
-      this.hostElement = document.getElementById(hostElementId)! as T;
-
-      const importedNode = document.importNode(this.templateElement.content, true);
-      this.element = importedNode.firstElementChild as U;
-      if (newElementId) {
-        this.element.id = newElementId;
-      }
-
-      this.attach(insertAtStart);
-    }
-
-    private attach(insertAtStart: boolean) {
-      this.hostElement.insertAdjacentElement(insertAtStart ? 'afterbegin' : 'beforeend', this.element);
-    }
-
-    abstract configure(): void;
-    abstract renderContent(): void;
-
-  }
-
   // ProjectItem class
   class ProjectItem extends Component<HTMLUListElement, HTMLLIElement> implements Draggable {
     private project: Project;
